Rename "metric" -> "metrics"
For constency with, e.g. django.db.models, rest_framework.serializers, etc.
